@@ -1,7 +1,7 @@
-import fs from "fs/promises";
+import fs from 'node:fs/promises'
 
-const cwd = process.cwd();
+const cwd = process.cwd()
 await Promise.all([
   fs.rm(`${cwd}/dist`, { recursive: true }).catch(() => {}),
   fs.rm(`${cwd}/type`, { recursive: true }).catch(() => {}),
-]);
+])
