@@ -1,16 +1,16 @@
-import { defineConfig } from 'rolldown'
 import terser from '@rollup/plugin-terser'
+import { defineConfig } from 'rolldown'
 
 const input = 'src/index.ts'
 const name = 'SaLambda'
 
 export default defineConfig([
   // CJS
-  { input, output:{ file:'dist/cjs/index.js', format:'cjs' } },
+  { input, output: { file: 'dist/cjs/index.js', format: 'cjs' } },
   // ESM
-  { input, output:{ file:'dist/esm/index.mjs', format:'esm' } },
+  { input, output: { file: 'dist/esm/index.mjs', format: 'esm' } },
   // UMD / IIFE main
-  { input, output:{ file:'dist/lib/sa-lambda.js', format:'iife', name } },
+  { input, output: { file: 'dist/lib/sa-lambda.js', format: 'iife', name } },
   // browser min
-  { input, output:{ file:'dist/lib/sa-lambda.min.js', format:'iife', name }, plugins: [terser()] }
+  { input, output: { file: 'dist/lib/sa-lambda.min.js', format: 'iife', name }, plugins: [terser()] },
 ])
