@@ -7,8 +7,8 @@
  * bimap :: Bifunctor f => f a c ~> (a -> b, c -> d) -> f b d
  */
 
-import { Functor2 } from './Functor'
-import { KindOf, URIS } from './HKT'
+import type { Functor2 } from './Functor'
+import type { KindOf, URIS } from './HKT'
 
 export interface Bifunctor2<F extends URIS> extends Functor2<F> {
   readonly bimap: <E, A, G, B>(fea: KindOf<F, [E, A]>, f: (e: E) => G, g: (a: A) => B) => KindOf<F, [G, B]>
