@@ -8,12 +8,12 @@
 export type TupleTail<T extends any[]> = T extends [any, ...infer R] ? R : never
 
 /**
-  * Returns last of tuple.
-  *
-  * ```ts
-  * TupleLast<[1, 2, 3]> => 3
-  * ```
-  */
+ * Returns last of tuple.
+ *
+ * ```ts
+ * TupleLast<[1, 2, 3]> => 3
+ * ```
+ */
 export type TupleLast<T extends any[]> = T[TupleTail<T>['length']]
 
 /**
