@@ -1,26 +1,24 @@
 import * as delay from './Delay'
+import * as effect from './Effect'
 import * as either from './Either'
 import * as equal from './Equal'
 import * as iterator from './Iterator'
 import * as math from './Math'
 import * as maybe from './Maybe'
-import * as effect from './Effect'
-import * as Tree from './Tree'
 
 export * from './Delay'
+export * from './Effect'
 export * from './Equal'
 export * from './function'
 export * from './Math'
 export * from './Pipe'
-export * from './Effect'
 export * from './Tree'
 export {
   delay,
+  effect,
   either,
   equal,
+  iterator,
   math,
   maybe,
-  effect,
-  iterator
 }
-
