@@ -1,10 +1,10 @@
-import { collectTreeNodes, filterTreeNodes } from '../src/Tree'
+import { expect, it } from 'vitest'
+import { collect, map } from '../src/Iterator'
 import { pipe } from '../src/Pipe'
-import { map, collect } from '../src/Iterator'
-import { fromPredicate } from '../src/Either'
+import { collectTreeNodes, filterTreeNodes } from '../src/Tree'
 
-test('collect tree nodes', () => {
-  type Node = {
+it('collect tree nodes', () => {
+  interface Node {
     id: number
     children?: Node[]
   }
@@ -12,24 +12,24 @@ test('collect tree nodes', () => {
   const root: Node = {
     id: 1,
     children: [{
-      id: 2
+      id: 2,
     }, {
       id: 3,
       children: [{
-        id: 4
-      }]
-    }]
+        id: 4,
+      }],
+    }],
   }
 
   expect(pipe(
-    collectTreeNodes(root, (a) => a.id === 2 || a.id === 4),
+    collectTreeNodes(root, a => a.id === 2 || a.id === 4),
     map(a => a.id),
-    collect
+    collect,
   )).toEqual([2, 4])
 })
 
-test('filter tree', () => {
-  type Node = {
+it('filter tree', () => {
+  interface Node {
     id: number
     children?: Node[]
   }
@@ -37,32 +37,32 @@ test('filter tree', () => {
   const root: Node = {
     id: 1,
     children: [{
-      id: 2
+      id: 2,
     }, {
       id: 3,
       children: [{
-        id: 4
-      }]
+        id: 4,
+      }],
     }, {
-      id: 5
-    }]
+      id: 5,
+    }],
   }
 
   expect(
-    filterTreeNodes(root, (a) => a.id === 1 || a.id >= 3),
+    filterTreeNodes(root, a => a.id === 1 || a.id >= 3),
   ).toEqual({
     id: 1,
     children: [{
       id: 3,
       children: [{
-        id: 4
-      }]
+        id: 4,
+      }],
     }, {
-      id: 5
-    }]
+      id: 5,
+    }],
   })
 
   expect(
-    filterTreeNodes(root, (a) => a.id === 2),
+    filterTreeNodes(root, a => a.id === 2),
   ).toEqual(null)
 })
